+++ CHANGELOG.md
@@ -10,3 +10,5 @@
 ## [0.1.0] - 2025-07-24
 
 - tests for version 0.1.0 stable
+- noDI, root, file functions
+- tests for noDI, root, file