+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 - tests for version 0.1.0 stable
 - noDI, root, file functions
 - tests for noDI, root, file
+- readme